Pair 클래스를 외부에서 사용하기 위해  HttpRequestUtils 클래스에서 외부로 꺼냄
